Resolve issues with example deployment
