+++ fltk/core/distributed/__init__.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from fltk.core.distributed.client import DistClient
 from fltk.core.distributed.dist_node import DistNode
 from fltk.core.distributed.extractor import download_datasets
