+++ requirements-cpu.txt
@@ -21,7 +21,7 @@ iteration-utilities==0.11.0
 Jinja2==3.1.1
 joblib==1.0.1
 kubeflow-training==1.5.0
-kubernetes==17.17.0
+kubernetes>=23.6.0
 Markdown==3.3.4
 MarkupSafe==2.1.1
 marshmallow==3.13.0
