+++ terraform/terraform-dependencies/main.tf
@@ -1,10 +1,7 @@
 
 data "google_client_config" "default" {}
 
-data "google_container_cluster" "testbed_cluster" {
-  name     = var.cluster_name
-  location = var.project_region
-}
+
 
 data "kustomization_build" "training-operator" {
   path = "github.com/kubeflow/manifests.git/apps/training-operator/upstream/overlays/standalone?ref=${var.kubeflow_version}"
@@ -30,7 +27,7 @@ resource "helm_release" "nfs_client_provisioner" {
       nfs_server_path  = var.nfs_provider_information.server_path
       image_repository = var.nfs_provider_information.image_repository
       image_tag        = var.nfs_provider_information.image_tag
-      image_policy     = var.nfs_provider_information.pull_policy
+      pull_policy     = var.nfs_provider_information.pull_policy
       nfs_size         = var.nfs_provider_information.storage_size
     })
   ]
