+++ terraform/terraform-dependencies/providers.tf
@@ -32,21 +32,25 @@ provider "kubectl" {
   config_path = var.kubernetes_config_path
 }
 
+
+data "google_container_cluster" "testbed_cluster" {
+  project  = var.project_id
+  name     = var.cluster_name
+  location = var.project_zone
+}
+
 provider "kubernetes" {
   host  = "https://${data.google_container_cluster.testbed_cluster.endpoint}"
-  token = data.google_client_config.provider.access_token # Provided by Google data object
+  token = data.google_client_config.default.access_token # Provided by Google data object
   cluster_ca_certificate = base64decode(
-    data.google_container_cluster.fltk_cluster.master_auth[0].cluster_ca_certificate,
+    data.google_container_cluster.testbed_cluster.master_auth[0].cluster_ca_certificate,
   )
 }
 
 
 provider "helm" {
   kubernetes {
-    host  = "https://${data.google_container_cluster.testbed_cluster.endpoint}"
-    token = data.google_client_config.provider.access_token # Provided by Google data object
-    cluster_ca_certificate = base64decode(
-      data.google_container_cluster.fltk_cluster.master_auth[0].cluster_ca_certificate,
-    )
+    config_path = var.kubernetes_config_path
   }
+
 }
