--- terraform/terraform-gke/resources.tf
@@ -1,13 +1,0 @@
-
-terraform {
-  required_providers {
-    # Interact with the project
-    google = {
-      source = "hashicorp/google"
-    }
-    kubernetes = {
-      source = "hashicorp/kubernetes"
-    }
-  }
-  required_version = ">= 1.0"
-}
