+++ terraform/terraform-gke/versions.tf
@@ -1,0 +1,13 @@
+
+terraform {
+  required_providers {
+    # Interact with the project
+    google = {
+      source = "hashicorp/google"
+    }
+    kubernetes = {
+      source = "hashicorp/kubernetes"
+    }
+  }
+  required_version = ">= 1.0"
+}
